+++ src/main/java/mil/nga/geopackage/tiles/features/FeatureTileGenerator.java
@@ -1,7 +1,9 @@
 package mil.nga.geopackage.tiles.features;
 
+import mil.nga.geopackage.BoundingBox;
 import mil.nga.geopackage.GeoPackage;
 import mil.nga.geopackage.extension.link.FeatureTileTableLinker;
+import mil.nga.geopackage.projection.Projection;
 import mil.nga.geopackage.tiles.TileGenerator;
 
 /**
@@ -26,14 +28,25 @@ public class FeatureTileGenerator extends TileGenerator {
 	 * Constructor
 	 *
 	 * @param geoPackage
+	 *            GeoPackage
 	 * @param tableName
+	 *            table name
 	 * @param featureTiles
+	 *            feature tiles
 	 * @param minZoom
+	 *            min zoom
 	 * @param maxZoom
+	 *            max zoom
+	 * @param boundingBox
+	 *            tiles bounding box
+	 * @param projection
+	 *            tiles projection
+	 * @since 1.2.0
 	 */
 	public FeatureTileGenerator(GeoPackage geoPackage, String tableName,
-			FeatureTiles featureTiles, int minZoom, int maxZoom) {
-		super(geoPackage, tableName, minZoom, maxZoom);
+			FeatureTiles featureTiles, int minZoom, int maxZoom,
+			BoundingBox boundingBox, Projection projection) {
+		super(geoPackage, tableName, minZoom, maxZoom, boundingBox, projection);
 		this.featureTiles = featureTiles;
 	}
 
@@ -64,12 +77,15 @@ public class FeatureTileGenerator extends TileGenerator {
 	@Override
 	protected void preTileGeneration() {
 
-		// Link the feature and tile table
-		if (linkTables) {
+		// Link the feature and tile table if they are in the same GeoPackage
+		GeoPackage geoPackage = getGeoPackage();
+		String featureTable = featureTiles.getFeatureDao().getTableName();
+		String tileTable = getTableName();
+		if (linkTables && geoPackage.isFeatureTable(featureTable)
+				&& geoPackage.isTileTable(tileTable)) {
 			FeatureTileTableLinker linker = new FeatureTileTableLinker(
-					getGeoPackage());
-			linker.link(featureTiles.getFeatureDao().getTableName(),
-					getTableName());
+					geoPackage);
+			linker.link(featureTable, tileTable);
 		}
 
 	}