+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Adheres to [Semantic Versioning](http://semver.org/).
 * Tile Creator providing common tile generation functionality
 * Removal of TileDraw, replaced by Tile Retriever and Tile Creator
 * Tile DAO changed to work with any projection units
+* Tile Generator support for multiple projections, such as WGS84 in addition to Web Mercator
+* URL Tile Generator changed to use provided projection in place of parsing URL
 
 ## [1.1.9](https://github.com/ngageoint/geopackage-java/releases/tag/1.1.9) (05-10-2016)